Introduce ResourceViewerViewModel for ResourceViewerFragment
- Created ResourceViewerViewModel encapsulating business logic (URL validation, updater creation).
- Injected required dependencies (authSessionUpdaterFactory, serverUrlMapper, sharedPrefManager, dispatcherProvider) directly into the ViewModel.
- Refactored ResourceViewerFragment to remove migrated logic and use the ViewModel via `by viewModels()`.
- Maintained Context-bound dependencies (like TTSManager) safely within the Fragment to prevent lifecycle leaks.
- Removed unused UserSessionManager dead code.
